refactor to start use commands
