+++ cmd/root.go
@@ -1,0 +1,104 @@
+package cmd
+
+import (
+	"fmt"
+	"net/http"
+	"os"
+
+	"github.com/auth0/go-jwt-middleware"
+	"github.com/dgrijalva/jwt-go"
+	"github.com/gorilla/mux"
+	"github.com/nuveo/prest/config"
+	"github.com/nuveo/prest/controllers"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+	"github.com/urfave/negroni"
+)
+
+var cfgFile string
+
+// RootCmd represents the base command when called without any subcommands
+var RootCmd = &cobra.Command{
+	Use:   "prest",
+	Short: "Serve a RESTful API from any PostgreSQL database",
+	Long:  `Serve a RESTful API from any PostgreSQL database, start HTTP server`,
+	Run: func(cmd *cobra.Command, args []string) {
+		app()
+	},
+}
+
+// Execute adds all child commands to the root command sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	if err := RootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(-1)
+	}
+}
+
+func init() {
+	cobra.OnInitialize(initConfig)
+
+	// Here you will define your flags and configuration settings.
+	// Cobra supports Persistent Flags, which, if defined here,
+	// will be global for your application.
+
+	RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.prest.yaml)")
+	// Cobra also supports local flags, which will only run
+	// when this action is called directly.
+	RootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
+
+// initConfig reads in config file and ENV variables if set.
+func initConfig() {
+	if cfgFile != "" { // enable ability to specify config file via flag
+		viper.SetConfigFile(cfgFile)
+	}
+
+	viper.SetConfigName(".prest") // name of config file (without extension)
+	viper.AddConfigPath("$HOME")  // adding home directory as first search path
+	viper.AutomaticEnv()          // read in environment variables that match
+
+	// If a config file is found, read it in.
+	if err := viper.ReadInConfig(); err == nil {
+		fmt.Println("Using config file:", viper.ConfigFileUsed())
+	}
+}
+
+func app() {
+	cfg := config.Prest{}
+	config.Parse(&cfg)
+
+	n := negroni.Classic()
+	n.Use(negroni.HandlerFunc(handlerSet))
+	if cfg.JWTKey != "" {
+		n.Use(jwtMiddleware(cfg.JWTKey))
+	}
+	r := mux.NewRouter()
+	r.HandleFunc("/databases", controllers.GetDatabases).Methods("GET")
+	r.HandleFunc("/schemas", controllers.GetSchemas).Methods("GET")
+	r.HandleFunc("/tables", controllers.GetTables).Methods("GET")
+	r.HandleFunc("/{database}/{schema}", controllers.GetTablesByDatabaseAndSchema).Methods("GET")
+	r.HandleFunc("/{database}/{schema}/{table}", controllers.SelectFromTables).Methods("GET")
+	r.HandleFunc("/{database}/{schema}/{table}", controllers.InsertInTables).Methods("POST")
+	r.HandleFunc("/{database}/{schema}/{table}", controllers.DeleteFromTable).Methods("DELETE")
+	r.HandleFunc("/{database}/{schema}/{table}", controllers.UpdateTable).Methods("PUT", "PATCH")
+
+	n.UseHandler(r)
+	n.Run(fmt.Sprintf(":%v", cfg.HTTPPort))
+}
+
+func handlerSet(w http.ResponseWriter, r *http.Request, next http.HandlerFunc) {
+	w.Header().Set("Content-Type", "application/json")
+	next(w, r)
+}
+
+func jwtMiddleware(key string) negroni.Handler {
+	jwtMiddleware := jwtmiddleware.New(jwtmiddleware.Options{
+		ValidationKeyGetter: func(token *jwt.Token) (interface{}, error) {
+			return []byte(key), nil
+		},
+		SigningMethod: jwt.SigningMethodHS256,
+	})
+	return negroni.HandlerFunc(jwtMiddleware.HandlerWithNext)
+}
